Move `login` route code into `Header` component
There is no reason for this to be implemented in a dedicated route that needs to use `transition.abort()` to avoid users from actually entering it. Directly running the code without the route transition seems significantly easier.
